Feat: Add getUserIdFromToken in TeamController

+++ build.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# set -e
+set -a
 build_part="server"
 version="${build_part}-v1.0"
 docker_image_name="ghcr.io/jujemu/match-up"
@@ -15,13 +15,16 @@ docker build --platform linux/amd64 -t ${docker_image_name}:${docker_version} .
 docker build -t "${build_part}":test .
 docker tag ${docker_image_name}:${docker_version} ${docker_image_name}:${build_part}
 
-echo "\n"
-echo -e "\033[1;33m # Docker image push on ghcr.io/jujemu\033[0m"
-docker push ${docker_image_name}:${docker_version}
-docker push ${docker_image_name}:${build_part}
-docker rmi ${docker_image_name}:${docker_version}
+if [ ${PUSH} = "push" ]
+then
+  echo "\n"
+  echo -e "\033[1;33m # Docker image push on ghcr.io/jujemu\033[0m"
+  docker push ${docker_image_name}:${docker_version}
+  docker push ${docker_image_name}:${build_part}
+  docker rmi ${docker_image_name}:${docker_version}
+fi
 
-if [ "$1" = "test" ]
+if [ "${TEST}" = "test" ]
 then
     echo "\n"
     echo -e "\033[1;33m # Docker run image built for test\033[0m"
